move customize method
